Moves TaskForm to its own Modal component
